fixed all register contrller bugs

+++ src/utils/cloudinary.js
@@ -1,4 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
+import dotenv from 'dotenv';
+dotenv.config();
 import fs from 'fs';
 
 cloudinary.config({
@@ -12,21 +14,28 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     try {
         // no file
-        if(!localFilePath) return null
+        if (!localFilePath) return null
 
         // upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type : 'auto'
+            resource_type: 'auto'
         })
         // file has been uploaded successflly
-        console.log('File Has Been Uploaded Successfully', response.url);
+        // console.log('File Has Been Uploaded Successfully', response.url);
+
+        // Delete file only after the file got uploaded in cloudinary
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return response;
-        
+
     } catch (error) {
         // removing the locally saved temporary file, as the upload operation gets failed
-        fs.unlinkSync(localFilePath)
+        fs.unlinkSync(localFilePath);
+        // console.log("Cloudinary Upload Error", error);
         return null
     }
 }
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
